fix(calibration): only load required weather data for ThermalCalibration
Move loading the weather data from the setup method of ThermalCalibration
to the process method.  Load only the chime weather data that is needed
to correct the input sidereal stream, rather than the entire range of
weather data covered by the calibration time file, to reduce the risk of
encountering an IO/filesystem errors.