feat(telescope): update beam evaluation for use in beamformer code
- Add angpos as a keyword argument to the beam method of the CHIME telescope.
Enables evaluation of the beam along a single source track, which is required
for the beamformer.  If not provided, then it will default to the _angos
telescope attribute.

- Add dec_normalized config parameter to CHIME class that, if set, will
normalize the beam by its magnitude at the transit of the declination
provided. To mach CHIME calibration, the user would set this parameter
to the declination of Cyg A.

- Creates a CHIMEParameterizedBeam class that is a CHIME telescope whose
beam method evaluates Mateus' parameterized fit to the driftscan beam.