feat(inject): tasks for injecting catalogs into visibility data
Introduces the FluxCatalog and SpectroscopicCatalog tasks,
which generate synthetic visibilities from a catalog of sources.
Also introduces versions of these tasks that modulate the signal
from each source by a beam model, either an external model, or
one implemented via the beam method of the provided telescope object.
The latter requires code in the ss/tel-beam branch.

Introduces code for perturbing the redshifts of the eBOSS tracers.