fix(sidereal): first check if source is present in FluxCatalog
The SiderealMean task provides the option to mask all sources
above some flux threshold.  However, there are now sources
that are present in ephemeris.source_dictionary but not in the
default collections loaded by the FluxCatalog, namely pulsars.
This fix first checks if a source is present in the FluxCatalog
before attempting to query it's flux.