config(daily): change low-weight masking
This replaces ThresholdVisWeightTime with a new version of
ThresholdVisWeightBaseline, which compares weights to a
median over freq and time for each baseline. The thresholds have
been tuned to catch obviously bad weights we found in chimestack
files by visual inspection, without flagging too much else.
It can happen that many cross-pol weights are low but co-pol
are not, so the pols_to_flag option implies that only co-pol weights are checked.

The previous version of ThresholdVisWeightBaseline is still
applied after regridding, under the new name
ThresholdVisWeightBaselineAlt.