feat(processing): add new chp `delay` processing type.
This adds a `delay` processing type to the types available in chp. The
basic implementation of the new type loads a sidereal stream file, runs
the standard daily processing masking, and produces a delay spectrum and
high-pass filtered delay spectrum. A given revision of the `delay` type
is linked to an existing revision of the `daily` type, so available tags
are pulled from the completed tags of the linked `daily` revision.