feat(SolveSources, solve_single_time): add source regularizer
At some times and frequencies, the fringe factors for multiple sources can
be very similar, such that the solutions for their amplitudes are not
very well distinguished. This commit adds an option to SolveSources, and
associated functionality to solve_single_time(), that looks for this
situation and interpolates the source solution from nearby time/freq samples
instead of solving directly.