feat(calibration): tasks for reconstructing/correcting errors in the gains
Introduces several new tasks.  ReconstructGainError calculates the ratio
of the gain to a low-pass filtered version of the gain, which can be
used as an estimate of the error in the gain.  In the process, it
iteratively flags frequency channels where the gain has been corrupted
by RFI spikes.  CorrectGainError applies the inverse of the gain error
to the stacked visibilites in an attempt to remove the errors that were
introduced by the real-time pipeline.  CollapseGainError simply averages
the gain error over all pairs of feeds.