ss/sidereal mean (#276)
* feat(sidereal): update sidereal mean tasks to handle other containers

Support for estimating and subtracting the mean or median over
sidereal day of both HybridVisStream and RingMap containers.

* feat(flagging): support for flagging HybridVisStream and RingMap

Updates DataFlagger and FlagRainfall to work with other
container types.

* feat(calibration): tasks for reconstructing/correcting errors in the gains

Introduces several new tasks.  ReconstructGainError calculates the ratio
of the gain to a low-pass filtered version of the gain, which can be
used as an estimate of the error in the gain.  In the process, it
iteratively flags frequency channels where the gain has been corrupted
by RFI spikes.  CorrectGainError applies the inverse of the gain error
to the stacked visibilites in an attempt to remove the errors that were
introduced by the real-time pipeline.  CollapseGainError simply averages
the gain error over all pairs of feeds.

* fix(flagging): fix daytime_flag to work with unsorted arrays

* feat(flagging): new flagging tasks

- MaskManyBadInputs: mask times where a large number of inputs were
  identifed as bad by the realtime system.
- MaskHighFracLost:  mask times and frequencies where a large fraction
  of data was lost in the realtime system (due to packetloss of RFI).
- MaskBrightSourcePixels: mask regions around the transit of bright
  sources in a ringmap.
- MaskBrightSourceTracks: mask regions around the U-shaped tracks
  of bright sources in a ringmap.
- MaskAliasedMap: mask regions of a map that suffer from aliasing.

* fix(calibration): capability to concatenate multiple GainData

Update ConcatenateGains so that it can handle GainData in addition
to StaticGainData.  Also redistribute over freq at the end of
ReconstructGainData to avoid OOM errors when writing to disk while
distributed over the last axis.

* feat(sidereal): default quarter ra ranges for mean estimation

Option to use the default RA range for each quarter when estimating the
crosstalk with the SiderealMean task.